import
The holo repository used to contain all the binaries that make up Holo:

* holo
* holo-build
* holo-files
* holo-run-scripts
* holo-users-groups
* (holo-test)

I'm splitting this big monolithic repo into multiple small ones to allow
for individual releases of the various components. This repository only
contains the holo-users-groups plugin.

The history for the monolithic repo, up until shortly before the 0.10
release of each of these components, can be found at
https://github.com/holocm/holo-attic