Save version information to a file, instead passing it as a flag.
In general; I prefer this approach, as it means that things are correctly
rebuilt if the version changes.

However, the push for it being done now is that I don't believe that the
linkerflags-based approach can be made to work with the `go test`-produced
binaries that an upcoming commit will add.

Now, because forcing the file to be generated would break `go get`, the
linkerflags-based approach still works; it just isn't used by the Makefile.
A `go get` user will have an empty version string.