Add the ability to create coverage-testing tooled binaries.
At a high-level, how this is done is by defining a TestXXX() function that
calls `main()`; and the "testing" subsystem takes care of everything else.

There are a couple of gotchas with this.

The big gotcha is that we can't call os.Exit(), or the process will exit
before the coverage results have been written.  Unfortunately, we need to
be able to exit with non-zero codes.  So, instead of having our Test
function call `main()`, we add a layer of indirection:

	+func main() {
	+	os.Exit(Main())
	+}
	+
	-func main() {
	+func Main() int {

and have the Test function call Main() instead of main().  However, now we have
to be disciplined and make sure that we *return from Main()* rather than call
os.Exit().  Although, this change isn't actually present in this commit because
I'd previously backported it to fix the leaking cachedir issue... but before a
rebase it was part of this commit.

Another issue is that the "testing" package requires use of the "flag"
package, rather than other argument parsing.  But holo doesn't use "flag".
So, we bypass "flag"'s normal behavior by calling
`flag.CommandLine.Parse(args)` ourselves with the flags we want to pass to
the testing subsystem, which we get from the environment variable
HOLO_TEST_FLAGS.

The final issue is that before exiting, the "testing" subsystem prints a
message informing the user that all tests ran successfully.  This breaks
our binaries, as we care about what happens on stdout.  So, we simply close
os.Stdout after calling Main(), but before returning control to the testing
subsystem.

Oh, and there's one more thing not related to the testing itself: You have
to give `go test` a list of packages that you want it to keep track of
coverage for.  Normally we could used `github.com/holocm/holo/...` to track
everything, but `go` won't follow symlinks when expanding `...` wildcards;
so we use `go list` to enumerate the packages that are actually getting
compiled.