test: Include empty directories in the tree.
Formerly it hard-coded a list of directories in HOLO_RESOURCE_DIR to
create; for each of the official plugins.  Because this would require each
of them to be listed in the `expected-tree`, and because it places a burden
on testing for non-official plugins, it now inspects the testcase's holorc
to determine which directories it needs to create.