Religiously avoid using os.Exit().  Fixes https://github.com/holocm/holo/issues/20
This is backported from a commit adding support for coverage testing.  I'm
not pull-requesting that entire branch because it rearranges the sources a
bit and is kind of invasive.

Anyway, in order to do coverage testing, I had to adjust it such that
instead of calling os.Exit(), it returns a value from a Main function,
which is wrapped by the actual main:

	func main() {
		os.Exit(Main())
	}

And, "for free" this fixed the cachedir leaking, as os.Exit was bypassing
the defer statement that cleans it up.