holo-files/common.FileBuffer: Change how it works, also store metadata.
Instead of being a pointer, it is now a direct struct.

It stores the FileMode (filetype and permissions), as well as the owner;
and writes these when you call .Write().

Because the path/basePath distinction is very holo-files specific, and
"common" mostly isn't; I've opted to have it not take the basePath
argument, and instead have the caller simply change the .Path of the
returned buffer if they need to override it.  I think it's clearer this
way.