holo-files: Rename things to avoid duplicate names for concepts.
There were the "external" names that documentation talked about, then there
were the internal names that the code used.  That's silly, and confusing to
would-be contributors.

  entity <- targetFile
  resource <- repoEntry/repoFile

And the use of "target" internally has been normalized:
They are now consistently properties of an entity.

  target (unchanged)
  base <- target base
  provisioned <- provisioned target <- last provisioned

I did not rename the files in the tests, though the names are now
confusing to anyone without knowledge of holo-files' internals before
this commit.