make entity IDs look the same in productive and testing mode
Entity IDs for holo-files usually look like

    file:/etc/foo.conf

But in testing mode, entity IDs used to look like

    file:target/etc/foo.conf

That did not pose a problem before, but it makes building a testcase for
alpm-hook.sh harder for @LukeShu. So now entity IDs always use the
absolute path of the target file (possibly within the scope of the
HOLO_ROOT_DIR).