Use the coverage-testing tooled binaries for running the test suite.
This required adjusting tests 11-fail-scan and 23-alpm-hook, in addition
to the shared holorc (which was enough to adjust all of the other tests).

The final coverage report is placed at `test/cov.cov`, which is built from
separate coverage reports for each invocation at `test/cov/*.cov`.

An HTML report is generated from this at `test/cov.html`.