holo-files: applyNonOrphan: Front-load all of the reads.
Because there is now more metadata associated with a FileBuffer, we
can just read everything all at once, instead of having to deal with a
tangle of pre-flight metadata checks.

Now, this patch actually adds code.  There's still a tangle of checks
that I didn't simplify--but it's really easy to do so now.  But I
figured I should keep this changeset smaller.

This pulls out some (small-ish) pieces into separate
.Get{Base,NewBaseProvisioned,Current,Desired} methods.  I plan to
share these with the applyOrphan code.