fix invocation of plugin command for "info" operation
The required environment variables (HOLO_STATE_DIR etc.)
were not set when the plugin command was invoked with the "info"
argument. This emerged as a problem because holo-users-groups creates
`${HOLO_STATE_DIR}/{base,provisioned}`, but since HOLO_STATE_DIR was not
given, it created `./{base,provisioned}` instead, which is not wanted.