Makefile: fix weird dependence on build/
The obviates the need for a "prepare-build" target.

I noticed this because `make clean` entirely wipes out `build/` including
`build/.keep`, which is checked in to git.

I think this was correct in some previous version, and got messed up in
merging one of the other plugins in to the monorepo.  IDK, I didn't dig in
to it.