holo-files: Fix first provision when there's a .pacnew file (#40)
This fixes two bugs that trigger when we are both:
 1. first provisioning a file that Holo has never touched before, and
 2. a system-update has already left a new stock version (a `.pacnew`
    file in this test) because of user modification.

The first is that because of a copy-paste mistake, it writes the contents
of the .pacnew file to the final file; both discarding user changes without
`--force`, and leaving the wrong thing in `$HOLO_STATE_DIR/base`.

The second is that the check to see if we need `--force` didn't work
anyway; in this situation, it needed to compare against `base` instead of
`provisioned`, since we haven't provisioned the file before.

Because both bugs trigger in the same situation, the fixes to both bugs are
required to make the added test pass.