doc: Consistently say "resource" instead of "source"
There's no reason to have two names for the same thing; it's
confusing.

At the same time, in holo(8), clarify that "source" and "resource" are
synonyms; I figure that "source" is in the website docs or a blog post
or some other docs, or at least in the memory of existing users.