holo-files: Implement .patch files.
This ended up being a lot more work than I thought it would be... though
a good chunk of that was slowly realizing all of the ways that I could mess
up writing patch files (in the tests) by hand.  Apparently, when applying
`--git` style patches, GNU patch is pretty finicky about the "index" line,
which I had expected it to basically ignore.

BusyBox patch doesn't implement the `-d` flag, despite it being in POSIX
for as long as `patch` has (2001; before that patch was standardized in
XDG/SUS, before that merged with POSIX; but I don't have pre-merge copies
handy to see if they had the `-d` flag).  So use `cmd.Dir` instead of `-d`.

That doesn't mean that BusyBox patch will pass the tests, just that it will
work.  It will still fail the tests:
 - The tests check for `--git` style changing of symlinks.  BusyBox patch
   does not support patching symlinks
 - The tests check for `--git` style changing of and file permissions.
   BusyBox patch does not support that
 - The tests expect it to say "patching file FILENAME" when creating a
   file (GNU patch behavior); BusyBox patch says "creating FILENAME".
 - The BusyBox fuzz detector seems to be less forgiving than the GNU one;
   it rejects my with-fuzz test.
That said, BusyBox patch will still work for users writing ordinary patch
files.

I will 100% not be surprised of the tests don't pass with BSD patch for
similar reasons.

This breaks backward compatibility in a very minor way: previously, a
resource file with the ".patch" suffix was understood to be a verbatim
file; this changes that, obviously.

The target file is operated on in a temporary directory, and without
specifying the `-p` flag or the target filename.  This gives the patch the
freedom to do directory-type operations (change file type).  If the patch
creates other files, they will be ignored.