Makefile: Unify lists of Go pkgs/dirs/srcs; exclude ./vendor from them
Everywhere where we needed a list of go packages or directories or
sources, there was an expression right there to find it.  Pull those
into variables.  This avoids overly long lines, and discrepancies
between lists.

That is, unify the lists of:
 - things included in coverage testing
 - things included in gofmt checking
 - things included in golint checking

In the above description, I mentioned "discrepencies".  The coverage
testing list included ./vendor, but the other lists didn't.  This
commit removes ./vendor from being included in coverage testing.  I
went with excluding it instead of including it because the toml
library does not pass golint.

This won't affect the coverage numbers reported by Coveralls because
the github.com/mattn/goveralls client that we use has its own logic to
filter out /vendor/ directories that we accidentally try to submit.