holo-test: make sure that GIT_DIR is unset for `git init`
GIT_DIR tells `git init` a different directory to use for ".git".  This
takes precedence over anything given as an argument.  This means that if
GIT_DIR is set, holo-test's `git init` invocation *won't* create a
temporary git repo in the current directory.

This will happen when using `git bisect` or the "exec" command of
`git rebase -i`, which will set GIT_DIR to the git repo being operated on.

That's mostly fine, as it's just a hack to work around a bug in the old
version of git on Travis; it's OK if on developers' boxes we accidentally
re-initialize the main holo git repo.  Except that, with tests now running in
parallel, every test will be trying to re-initialize the same git repo,
maybe at the same time.  Because of locking, some of them will fail.

That's also mostly fine, but will spit out a scary-looking message on
stderr:

    error: could not lock config file /home/lukeshu/src/holocm/holo/.git/config: File exists
    fatal: could not set 'core.repositoryformatversion' to '0'

And it's really scary when your screen fills up with 60 lines of that
unexpectedly. :)