extend generator environment, fix symlink calculation
This commit started out with me adding $HOLO_RESOURCE_ROOT and
$HOLO_CACHE_DIR to the generator environment, and putting both $OUT and
$HOLO_CACHE_DIR for the generator below the CacheDirectory().

I then had to adjust the testcases because $HOLO_RESOURCE_DIR was now
indeterministic (since it resides in CacheDirectory() whose name is
randomly generated) and thus `apply-output` became indeterminstic as
well.

I also had to extend the echo_plugin to dump the contents of
$HOLO_RESOURCE_DIR: Since those were in CacheDirectory(), they are
cleaned up at the end of each Holo run and thus do not show up in
`tree`.

While doing all of this, I noticed that the echo_plugin does
not actually cover access to the generated resource files. Upon
extending the echo_plugin to actually print the contents of each file, I
found that the symlinks generated by symlinkFiles() are actually broken,
at least in test runs when RootDirectory() is not `/`, so I fixed that
as well.