make the virtualized resource directory mandatory even without generated resource files
This change sounds like needless complication at first, but I'm very
glad that I did it. The existing set of tests only tested generators
with a mock plugin, not with actual plugins like holo-files. Turns out,
holo-files goes to shit as soon as it doesn't get /usr/share/holo/files
anymore because it scans with filepath.Walk() which does not traverse
symlinks.

Therefore we have to copy static resource files into the virtual
resource directory instead of just symlinking directories. Even worse:
We cannot even symlink the individual files because files and symlinks
have different semantics in holo-files.

I always try to have the tests passing when making a commit, but this
commit is an exception because the changes required to fix the tests are
massive enough to warrant a separate commit. Right now, the tests fail
because scan reports now refer to the resource files in the virtualized
resource directory. Random example from the test output:

    - apply target/usr/share/holo/files/01-first/etc/symlink-to-file.conf
    + apply target/tmp/holo/generated-resources/files/01-first/etc/symlink-to-file.conf

And it's only that clean because we clean /tmp/holo.135918760 into
/tmp/holo in holo-test. (:

This is not even just a cosmetic error. CLI selectors use the source
file references from the scan report, so this breaks file selectors
entirely. We will have to do some magic to rewrite the virtual resource
files back into their physical selves.