do not copy all file permissions and ownership into virtualized resource directory
This fixes running `holo scan` as non-root. When I saw that bug and dug
into it, I had a minor heart attack upon seeing the root cause, because
I thought that fixing this was going to break holo-files. My heart
unattacked when I found that holo-files does not honor file permissions
and ownership on resource files at all (except for needing the
executable bit on holoscripts). At least for the core plugins, this is a
safe change, but I noted it down in the CHANGELOG as a potentially
disruptive difference in behavior.